+++ src/Model/LeadTrackerModel.cs
@@ -1,0 +1,7 @@
+﻿namespace NBASharp.Model;
+
+public class LeadTrackerModel
+{
+    public InternalModel Internal { get; set; }
+    public List<PlayModel> Plays { get; set; }
+}