+++ src/Model/PlayModel.cs
@@ -1,0 +1,8 @@
+﻿namespace NBASharp.Model;
+
+public class PlayModel
+{
+    public string Clock { get; set; }
+    public string LeadTeamId { get; set; }
+    public string Points { get; set; }
+}