+++ src/ModelDto/LeadTrackerRootDto.cs
@@ -1,0 +1,7 @@
+﻿namespace NBASharp.ModelDto;
+
+public class LeadTrackerRootDto
+{
+    public InternalDto _internal { get; set; }
+    public List<PlayDto> plays { get; set; }
+}