+++ src/ModelDto/PlayDto.cs
@@ -1,0 +1,8 @@
+﻿namespace NBASharp.ModelDto;
+
+public class PlayDto
+{
+    public string clock { get; set; }
+    public string leadTeamId { get; set; }
+    public string points { get; set; }
+}